scripts/sct_analyze_lesion: move analyze_lesion to process_seg
